Add files via upload
The entire project works on a network setup between a mobile camera, a laptop and a node MCU. Image from the camera is fed to the laptop via hot spot and this image is processed using OpenCV library in python so as to find the centroid of a particular colored object. Then, with the application of pid concept, the required motion of the bot is calculated so as to make the bot follow the object. This calculated value is then transfered to the Node MCU placed on the bot which then gives instructions to the bot, and the bot follows the object accordingly. This repository includes the python code "ip traking pid.py" and the arduino code "ip_obj_tracking_pid.ino".